refactor: Refactoring of Build to separate SiteBuilder
